generic/py/pyproject.toml
[project]
name = "generic"
version = "0.1.0"
description = "generic -- a place to think"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "sqlalchemy[asyncio]",
    "psycopg2-binary",
    "asyncpg",
    "alembic",

    "python-dotenv",
    "python-multipart",
    "pydantic",
    "tenacity",
    
    "uvicorn",
    "fastapi",
    "fastapi-sso",
    "fastapi-login",
    
    "greenlet",
    "starlette",
    "sse-starlette",
    
    "jinja2",
    
    "python-jose[cryptography]",
    "types-python-jose",
    "itsdangerous==2.1.2",
    "werkzeug==3.0.1",

    "setuptools",
    "watchfiles",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.25.3",
    "alembic",
]
